requirements.txt
absl-py==2.1.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
boto3==1.34.131
botocore==1.34.131
cargo==0.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==4.2.6
django-deep-serializer==0.1.3
et-xmlfile==1.1.0
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.13.1
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.44.0
fqdn==1.5.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
gunicorn==21.2.0
huggingface-hub==0.23.4
idna==3.7
imbalanced-learn==0.12.0
imblearn==0.0
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.29
jaxlib==0.4.29
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lapx==0.5.8
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
opencv-contrib-python==4.10.0.82
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytweening==1.2.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.33
rpds-py==0.10.6
s3transfer==0.10.1
sahi==0.11.16
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==70.1.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.7
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
style==1.1.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tornado==6.3.3
tqdm==4.66.2
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.2.6
update==0.0.1
uri-template==1.3.0
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
yolov5==7.0.13
yolov8==0.0.2
